Improve fake modifier keys
They can now be either turned on for one keystroke, or latched on
for all following keystrokes.  A primitive UI displays the status
of the modifiers before each keystroke.  The command key combo is
switched to Ctrl-q, since Ctrl-b is used for marking text as bold.
A verbose mode will print out keystrokes, but the default mode is
to show less output.